+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0] - 2025-02-03
+
+### Changed
+
+- `Order.isInsideGeofence` is now an async fuction that returns the value at the moment when it called (instead of the constant value at the time of `getOrders` being called)
+- Updated HyperTrack SDK iOS to [5.11.0](https://github.com/hypertrack/sdk-ios/releases/tag/5.11.0)
+- Updated HyperTrack SDK Android to [7.11.0](https://github.com/hypertrack/sdk-android/releases/tag/7.11.0)
+
 ## [3.8.0] - 2025-01-17
 
 ### Added
@@ -348,3 +356,4 @@ We are excited to announce the release of HyperTrack Ionic Capacitor SDK 2.0.0, 
 [3.7.0]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.7.0
 [3.7.1]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.7.1
 [3.8.0]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/3.8.0
+[4.0.0]: https://github.com/hypertrack/sdk-ionic-capacitor/releases/tag/4.0.0